rename post.go file
